backend/requirements.txt
asgiref==3.3.4
attrs==21.2.0
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-spectacular==0.17.0
inflection==0.5.1
jsonschema==3.2.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
